Remove legacy docs and consolidate deployment summary into main README
