+++ archive/docs-backup-$(date +%Y%m%d)/CLEANUP_SUMMARY.md
@@ -1,0 +1,55 @@
+# 文檔清理摘要
+
+## 🧹 清理日期
+$(date +%Y-%m-%d)
+
+## 📊 清理前後對比
+
+### 清理前
+- 19 個 .md 檔案
+- 包含重複和冗餘的文檔
+- 包含舊版本的程式碼檔案
+
+### 清理後
+- 16 個活躍的 .md 檔案
+- 整合和簡化的文檔結構
+- 移除舊版本檔案
+
+## 🗂️ 已歸檔的檔案
+
+### 1. 冗餘文檔
+- `DEPLOYMENT_SUMMARY.md` - 專案完成摘要，資訊已整合到其他文檔中
+
+### 2. 舊版程式碼
+- `main.py` (重新命名為 `legacy-main.py`) - 3,070+ 行的舊版單體程式，已被模組化架構取代
+- `migrate_to_modular.py` - 遷移測試腳本，遷移已完成
+
+## ✅ 保留和改進的檔案
+
+### 主要改進
+- `./wazuh-docker/single-node/README.md` - 已整合和簡化，提供清晰的部署指引
+
+### 保留的重要文檔
+- `UNIFIED_STACK_README.md` - 完整的統一堆疊部署指南
+- `MIGRATION_GUIDE.md` - 架構遷移指南（仍有參考價值）
+- AI agent 相關文檔（docs/ 目錄）
+- 各組件的 README.md 檔案
+
+## 🎯 清理成果
+
+1. **消除冗餘**: 移除重複和過時的文檔
+2. **簡化結構**: 整合相關資訊到主要文檔中
+3. **保持可追溯性**: 所有移除的檔案都已安全歸檔
+4. **改善可維護性**: 減少需要同步更新的文檔數量
+
+## 📝 建議
+
+1. **定期檢查**: 每季度檢查文檔是否需要更新
+2. **避免重複**: 新建文檔前檢查現有資源
+3. **版本控制**: 重要變更應該有版本記錄
+4. **整合原則**: 相關資訊盡量整合在同一文檔中
+
+---
+
+**備份位置**: `./archive/docs-backup-$(date +%Y%m%d)/`  
+**執行者**: AI 文檔清理助手
